+++ configs/ping/policy.yml
@@ -7,4 +7,3 @@ policies:
           - path: "/usr/sbin/ping"
         sockets:
           - protocol: "icmpv4"
-            remote_ip: "10.1.3.10"
